+++ experiment/measurer/coverage_utils.py
@@ -15,8 +15,6 @@
 
 import os
 import json
-import hashlib
-import pandas as pd
 
 from common import experiment_path as exp_path
 from common import experiment_utils as exp_utils
@@ -56,129 +54,6 @@ def generate_coverage_reports(experiment_config: dict):
             generate_coverage_report(experiment, benchmark, fuzzer)
 
     logger.info('Finished generating coverage reports.')
-
-
-class DataFrameContainer:  # pylint: disable=too-many-instance-attributes
-    """Maintains segment and function coverage information, and writes this
-    information to CSV files."""
-
-    def __init__(self):
-        """Constructor"""
-        self.segment_entries = []
-        self.function_entries = []
-        self.fuzzer_names = dict()
-        self.benchmark_names = dict()
-        self.function_names = dict()
-        self.file_names = dict()
-        # Will be initialized upon done_adding_entries().
-        self.segment_df = None
-        self.function_df = None
-        self.name_df = None
-
-    def add_function_entry(self, benchmark, fuzzer, trial_id, function,
-                           function_hits, time):
-        # pylint: disable=too-many-arguments
-        """Adds an entry to the function_df."""
-        fuzzer_id = self.name_to_id(fuzzer)
-        benchmark_id = self.name_to_id(benchmark)
-        function_id = self.name_to_id(function)
-        self.fuzzer_names[fuzzer] = fuzzer_id
-        self.benchmark_names[benchmark] = benchmark_id
-        self.function_names[function] = function_id
-
-        function_entry = [
-            benchmark_id, fuzzer_id, trial_id, time, function_id, function_hits
-        ]
-        self.function_entries.append(function_entry)
-
-    def add_segment_entry(self, benchmark, fuzzer, trial_id, file_name, line,
-                          column, time):
-        # pylint: disable=too-many-arguments
-        """Adds an entry to the segment_df."""
-        fuzzer_id = self.name_to_id(fuzzer)
-        benchmark_id = self.name_to_id(benchmark)
-        file_id = self.name_to_id(file_name)
-        self.fuzzer_names[fuzzer] = fuzzer_id
-        self.benchmark_names[benchmark] = benchmark_id
-        self.file_names[file_name] = file_id
-
-        segment_entry = [
-            benchmark_id, fuzzer_id, trial_id, time, file_id, line, column
-        ]
-        self.segment_entries.append(segment_entry)
-
-    def done_adding_entries(self):
-        """Generates the data frames from the individual entries."""
-
-        if len(self.segment_entries) == 0:
-            logger.error('Finalizing, but no entries were added.')
-            return
-
-        self.segment_df = pd.DataFrame(self.segment_entries,
-                                       columns=[
-                                           "benchmark", "fuzzer", "trial",
-                                           "time", "file", "line", "column"
-                                       ])
-        self.function_df = pd.DataFrame(self.function_entries,
-                                        columns=[
-                                            "benchmark", "fuzzer", "trial",
-                                            "time", "function", "hits"
-                                        ])
-
-        name_entries = []
-        for name in self.benchmark_names:
-            name_entries.append([self.benchmark_names[name], name, "benchmark"])
-        for name in self.fuzzer_names:
-            name_entries.append([self.fuzzer_names[name], name, "fuzzer"])
-        for name in self.function_names:
-            name_entries.append([self.function_names[name], name, "function"])
-        for name in self.file_names:
-            name_entries.append([self.file_names[name], name, "file"])
-        self.name_df = pd.DataFrame(name_entries,
-                                    columns=["id", "name", "type"])
-
-    def name_to_id(self, name):  # pylint: disable=no-self-use
-        """Generates a hash for the name. This is to save disk storage"""
-        return hashlib.md5(name.encode()).hexdigest()[:7]
-
-    def remove_redundant_duplicates(self):
-        """Removes redundant entries in segment_df. Before calling this
-        function, for each time stamp, segment_df contains all segments that are
-        covered in this time stamp. After calling this function, for each time
-        stamp, segment_df only contains segments that have been covered since
-        the previous time stamp. This significantly reduces the size of the
-        resulting CSV file."""
-        try:
-            # Drop duplicates but with different timestamps in segment data.
-            self.segment_df = self.segment_df.sort_values(by=['time'])
-            self.segment_df = self.segment_df.drop_duplicates(
-                subset=self.segment_df.columns.difference(['time']),
-                keep="first")
-            self.name_df = self.name_df.drop_duplicates(keep="first")
-
-        except (ValueError, KeyError, IndexError):
-            logger.error('Error occurred when removing duplicates.')
-
-    def generate_csv_files(self):
-        """Generates three compressed CSV files containing coverage information
-        for all fuzzers, benchmarks, and trials. To maintain a small file size,
-        all strings, such as file and function names, are referenced by id and
-        resolved in 'names.csv'."""
-
-        # Clean and prune experiment-specific data frames.
-        self.remove_redundant_duplicates()
-
-        # Write CSV files to file store.
-        def csv_filestore_helper(file_name, df):
-            """Helper function for storing csv files in filestore."""
-            src = os.path.join(get_coverage_info_dir(), 'data', file_name)
-            dst = exp_path.filestore(src)
-            df.to_csv(src, index=False, compression='infer')
-            filestore_utils.cp(src, dst)
-
-        csv_filestore_helper('functions.csv.gz', self.function_df)
-        csv_filestore_helper('segments.csv.gz', self.segment_df)
-        csv_filestore_helper('names.csv.gz', self.name_df)
 
 
 def generate_coverage_report(experiment, benchmark, fuzzer):
@@ -393,43 +268,6 @@ def generate_json_summary(coverage_binary,
                                      output_file=dst_file,
                                      expect_zero=False)
     return result
-
-
-def extract_segments_and_functions_from_summary_json(  # pylint: disable=too-many-locals
-        summary_json_file, benchmark, fuzzer, trial_id, time):
-    """Return a trial-specific data frame container with segment and function
-     coverage information given a trial-specific coverage summary json file."""
-
-    process_specific_df_container = DataFrameContainer()
-
-    try:
-        coverage_info = get_coverage_infomation(summary_json_file)
-        # Extract coverage information for functions.
-        for function_data in coverage_info['data'][0]['functions']:
-            process_specific_df_container.add_function_entry(
-                benchmark, fuzzer, trial_id, function_data['name'],
-                function_data['count'], time)
-
-        # Extract coverage information for segments.
-        for file in coverage_info['data'][0]['files']:
-            for segment in file['segments']:
-                if segment[2] != 0:  # Segment hits.
-                    process_specific_df_container.add_segment_entry(
-                        benchmark,
-                        fuzzer,
-                        trial_id,
-                        file['filename'],
-                        segment[0],  # Segment line.
-                        segment[1],  # Segment column.
-                        time)
-
-    except (ValueError, KeyError, IndexError):
-        logger.error('Failed when extracting trial-specific segment and'
-                     'function information from coverage summary')
-
-    process_specific_df_container.done_adding_entries()
-
-    return process_specific_df_container
 
 
 def extract_covered_regions_from_summary_json(summary_json_file):
